Made setup.py, relocated files
Having random repo cruft in the package is less than ideal.  The ideal solution is to have the package sit inside a subdirectory of the repository, so things like the readme and examples are not part of the package.

setup.py makes this installable

Made imports consistent by switching all of them to explicit full imports.

Removed an empty unimported module.
