--- __init__.py
@@ -1,8 +1,0 @@
-from .Component import Component
-from .Window import Window
-from .AutoProperty import AutoProperty
-from tkpf.Bindable import Bindable
-from .Binding import Binding
-from .ViewModel import ViewModel
-from .NumericEntry import NumericEntry
-from .Menu import Menu as _Menu
